chore: add compleation block for didFinalize
