+++ AdyenComponents/Apple Pay/ApplePayComponentError.swift
@@ -1,0 +1,62 @@
+//
+// Copyright (c) 2022 Adyen N.V.
+//
+// This file is open source and available under the MIT license. See the LICENSE file for more info.
+//
+
+import Adyen
+import Foundation
+import PassKit
+
+extension ApplePayComponent {
+
+    /// Describes the error that can occur during Apple Pay payment.
+    public enum Error: Swift.Error, LocalizedError {
+        /// Indicates that the user can't make payments on any of the payment request’s supported networks.
+        case userCannotMakePayment
+
+        /// Indicates that the current device's hardware doesn't support ApplePay.
+        case deviceDoesNotSupportApplyPay
+
+        /// Indicates that the summaryItems array is empty.
+        case emptySummaryItems
+
+        /// Indicates that the grand total summary item is a negative value.
+        case negativeGrandTotal
+
+        /// Indicates that at least one of the summary items has an invalid amount.
+        case invalidSummaryItem
+
+        /// Indicates that the country code is invalid.
+        case invalidCountryCode
+
+        /// Indicates that the currency code is invalid.
+        case invalidCurrencyCode
+
+        /// Indicates that the token was generated incorrectly.
+        case invalidToken
+
+        /// :nodoc:
+        public var errorDescription: String? {
+            switch self {
+            case .userCannotMakePayment:
+                return "The user can’t make payments on any of the payment request’s supported networks."
+            case .deviceDoesNotSupportApplyPay:
+                return "The current device's hardware doesn't support ApplePay."
+            case .emptySummaryItems:
+                return "The summaryItems array is empty."
+            case .negativeGrandTotal:
+                return "The grand total summary item should be greater than or equal to zero."
+            case .invalidSummaryItem:
+                return "At least one of the summary items has an invalid amount."
+            case .invalidCountryCode:
+                return "The country code is invalid."
+            case .invalidCurrencyCode:
+                return "The currency code is invalid."
+            case .invalidToken:
+                return "The Apple Pay token is invalid. Make sure you are using physical device, not a Simulator."
+            }
+        }
+    }
+
+}
