+++ AdyenDropIn/DropInComponentExtensions.swift
@@ -123,9 +123,9 @@ extension DropInComponent: FinalizableComponent {
     /// Stops loading and finalize DropIn's selected payment if necessary.
     /// This method must be called after certain payment methods (e.x. ApplePay)
     /// - Parameter success: Status of the payment.
-    public func didFinalize(with success: Bool) {
+    public func didFinalize(with success: Bool, compleate: (() -> Void)?) {
         stopLoading()
-        selectedPaymentComponent?.finalizeIfNeeded(with: success)
+        selectedPaymentComponent?.finalizeIfNeeded(with: success, compleate: compleate)
     }
 }
