+++ Demo/Configuration.swift
@@ -37,7 +37,7 @@ internal enum ConfigurationConstants {
 
     static let demoServerAPIKey = "{YOUR_DEMO_SERVER_API_KEY}"
 
-    static let applePayMerchantIdentifier = "{YOUR_APPLE_PAY_MERCHANT_IDENTIFIER}"
+    static let applePayMerchantIdentifier = "merchant.com.adyen.checkoutsandbox.test"
 
     static let merchantAccount = "{YOUR_MERCHANT_ACCOUNT}"
